+++ test/unittest/mock/file_io_mock.cpp
@@ -27,7 +27,7 @@ typedef off_t (*LseekFuncT)(int fd, off_t offset, int whence);
 typedef ssize_t (*WriteFuncT)(int fd, const void *buf, size_t count);
 typedef int (*FtruncateFuncT)(int fd, off_t length);
 typedef errno_t (*MemcpyFuncT)(void *dest, size_t destMax, const void *src, size_t count);
-typedef int (*ReadFuncT)(int fd, void *dest, uint32_t maxCount);
+typedef ssize_t (*ReadFuncT)(int fd, void *dest, size_t maxCount);
 
 off_t lseek(int fd, off_t offset, int whence)
 {
@@ -86,7 +86,7 @@ errno_t memcpy_s(void *dest, size_t destMax, const void *src, size_t count)
     return (*func)(dest, destMax, src, count);
 }
 
-ssize_t read(int fd, void *dest, uint32_t maxCount)
+ssize_t read(int fd, void *dest, size_t maxCount)
 {
     if (IsFuncNeedMock("read")) {
         CommonMockFuncT rawFunc = GetMockFunc(__func__);
